Support getting passwords asynchronously.
When using EXWM, if `password-store-get` is called and a pinentry
program needs to be executed, Emacs deadlocks.  This happens becuase
Emacs blocks waiting for output from `gpg(1)`, which is blocked
waiting for output from the pinentry-program, which is blocked waiting
for Emacs to manage its window.

This PR adds `password-store-get-async` (and supporting functions),
which spawn an async process to invoke gpg.  It also updates
`password-store-copy` to use the async method by default.

re NicolasPetton/pass#25